refactor: move modulos de diabetes para dentro de /modules
