+++ src/features/acf/backend/diabetes/modules/filters/repository.ts
@@ -1,5 +1,5 @@
 import { prisma } from "@prisma/prismaClient";
-import type { FilterOptionsDb } from "../model";
+import type { FilterOptionsDb } from "../../model";
 
 export const filterOptionsCoaps = async (
     municipalitySusId: string
