+++ src/features/acf/backend/diabetes/modules/internalCards/adapter.ts
