+++ src/features/acf/backend/diabetes/modules/table/adapter.ts
