+++ src/features/acf/backend/diabetes/modules/table/index.ts
@@ -1,0 +1,1 @@
+export { page, rowCount } from "./controller";
